Feat: 별마크 조회 코드 작성, User Application 응답 데이터 수정, 테스트코드 작성

See merge request s10-final/S10P31C105!78

+++ BackEnd/StellAR/src/main/java/com/ssafy/stellar/user/service/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.ssafy.stellar.user.service;
 
 import com.ssafy.stellar.user.dto.request.SignUpDto;
+import com.ssafy.stellar.user.dto.response.UserDto;
 import com.ssafy.stellar.user.entity.UserEntity;
 import com.ssafy.stellar.user.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
@@ -25,14 +26,23 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void signUp(SignUpDto signUpDto) {
+
         String encodePassword = bCryptPasswordEncoder.encode(signUpDto.getPassword());
-        signUpDto.setPassword(encodePassword);
-        UserEntity user = signUpDto.toEntity();
+        String userId = signUpDto.getUserId();
+
+        if (userRepository.findByUserId(userId) != null) {
+            throw new RuntimeException("User already exists with id: " + userId);
+        }
+
+        UserEntity user = new UserEntity();
+        user.setUserId(userId);
+        user.setPassword(encodePassword);
+
         userRepository.save(user);
     }
 
     @Override
-    public UserEntity logIn(String userId, String password) {
+    public UserDto logIn(String userId, String password) {
 
         UserEntity user = userRepository.findByUserId(userId);
         if (user == null) {
@@ -41,14 +51,24 @@ public class UserServiceImpl implements UserService {
         }
 
         if (checkPassword(userId, password)) {
-            return user;
+            UserDto userDto = new UserDto();
+            userDto.setUserId(user.getUserId());
+            userDto.setName(user.getName());
+            userDto.setGender(user.getGender());
+
+            return userDto;
         } else {
-            return null;
+            throw new IllegalArgumentException("Invalid password");
         }
     }
 
     @Override
     public void deleteUser(String userId) {
+
+        if (!userRepository.existsById(userId)) {
+            throw new UsernameNotFoundException("User not found with id: " + userId);
+        }
+
         userRepository.deleteById(userId);
     }
 
@@ -56,7 +76,6 @@ public class UserServiceImpl implements UserService {
     public boolean checkPassword(String userId, String password) {
         UserEntity user = userRepository.findByUserId(userId);
         if (user == null) {
-            // 로그인 실패 처리 로직 또는 예외 발생
             throw new UsernameNotFoundException("User not found with id: " + userId);
         }
         return bCryptPasswordEncoder.matches(password, user.getPassword());
